refactor: better modules for client
